requirements.txt
flask
flask-script==2.0.5
Flask-SQLAlchemy==2.2
psycopg2==2.7.1
Flask-Testing==0.6.2
coverage==4.4.1
flask-bcrypt==0.7.1
pyjwt==1.5.0